Completed Admin Service Bulk User Creation

+++ src/main/java/com/waya/wayaauthenticationservice/repository/RolesRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.waya.wayaauthenticationservice.entity.Roles;
 
 @Repository
-public interface RolesRepository extends JpaRepository<Roles, Integer> {
+public interface RolesRepository extends JpaRepository<Roles, Long> {
 	
 	 Optional<Roles> findByName(String name);
 }
